update bm25 script with prevent-rewrite flag

+++ src/5_bm25.py
@@ -1,3 +1,5 @@
+import os
+
 from utils import load_documents, split_documents, build_bm25_retriever, bm25_search, print_top_results
 import pickle
 import argparse
@@ -23,6 +25,10 @@ def parse_args():
                         default=100,
                         type=int,
                         help="Overlap between chunks")
+    parser.add_argument("--prevent-rewrite",
+                        type=bool, 
+                        default=True,
+                        help="Prevent rewriting the index if one exists")
     # For extra feature to test query after building index
     parser.add_argument("--query",
                         default=None,
@@ -37,27 +43,35 @@ def main():
     '''Main function to build a LangChain BM25 retriever from a parquet file of product documents
     and optionally tests it with a query.'''
     print("[STATUS] Starting BM25 index process")
-    args = parse_args()
-
-    # Document loading and splitting
-    print("[STATUS] Loading and chunking/splitting docs")
-    whole_docs = load_documents(parquet_path = args.input, 
-                                text_col= args.text_col)
-    split_docs = split_documents(whole_docs, 
-                                args.chunk_size,
-                                args.chunk_overlap)
-
-    #create Langchain BM25 retriever
-    print("[STATUS] Building BM25 index")
-    bm25_retriever = build_bm25_retriever(split_docs)
     
-    with open(f"{args.output_dir}langc_bm25_index.pkl", "wb") as f:
-        pickle.dump(bm25_retriever, f)
-    print(f"[SAVED] LangChain BM25 retriever as .pkl file")
+    args = parse_args()
+    index_path = f"{args.output_dir}bm25_index.pkl"
 
-    with open(f"{args.output_dir}corpus.pkl", "wb") as f:
-        pickle.dump(bm25_retriever.docs, f)
-    print("[SAVED] Saved BM25 sub-docs corpus")
+    if not args.prevent_rewrite or not os.path.exists(index_path):
+    # Document loading and splitting
+        print("[STATUS] Loading and chunking/splitting docs")
+        whole_docs = load_documents(parquet_path = args.input, 
+                                    text_col= args.text_col)
+        split_docs = split_documents(whole_docs, 
+                                    args.chunk_size,
+                                    args.chunk_overlap)
+
+        #create Langchain BM25 retriever
+        print("[STATUS] Building BM25 index")
+        bm25_retriever = build_bm25_retriever(split_docs)
+        
+        with open(f"{index_path}", "wb") as f:
+            pickle.dump(bm25_retriever, f)
+        print(f"[SAVED] LangChain BM25 retriever as .pkl file")
+
+        with open(f"{args.output_dir}corpus.pkl", "wb") as f:
+            pickle.dump(bm25_retriever.docs, f)
+        print("[SAVED] Saved BM25 sub-docs corpus")
+    else: 
+        print("[STATUS] Index already exists, loading index from file.")
+        with open(index_path, "rb") as f:
+            bm25_retriever = pickle.load(f)
+        print("[DONE] Index loaded")
     print("[STATUS] Finished BM25 process.")
     
     # Extra feature if user also passed in a test query
